chore: move validator higher in the folder structure and reuse project name max length
