+++ application/ui/src/features/project/create/create-project-form.tsx
@@ -13,12 +13,12 @@ import type { Label, Project, TaskType } from '../../../constants/shared-types';
 import { LabelSelection } from '../label-selection/label-selection.component';
 import { TASK_OPTIONS, TaskSelection } from '../task-selection/task-selection.component';
 import { isClassificationTask } from '../task-type-guards';
+import { PROJECT_NAME_MAX_LENGTH, validateProjectName } from '../validator';
 import {
     ClassificationTaskSelection,
     ClassificationTaskType,
 } from './classification-label-selection/classification-task-type-selection.component';
 import { generateUniqueProjectName } from './utils';
-import { validateProjectName } from './validator';
 
 import classes from './create-project-form.module.scss';
 
@@ -93,6 +93,7 @@ export const CreateProjectForm = ({ projects }: CreateProjectFormProps) => {
                 <Flex justifyContent={'center'} marginTop={'size-600'}>
                     <TextField
                         aria-label={'Project name input'}
+                        maxLength={PROJECT_NAME_MAX_LENGTH}
                         isRequired
                         value={name}
                         onChange={setName}
