+++ application/ui/src/features/project/list/import-dataset-as-new-project/import-task-selection/import-task-selection.component.tsx
@@ -10,8 +10,8 @@ import { useStagedDatasetSuspense } from 'hooks/api/staged-dataset.hook';
 import { useImportDatasetAsNewProject } from 'hooks/storage/use-import-dataset-as-new-project.hook';
 
 import { TaskType } from '../../../../../constants/shared-types';
-import { validateProjectName } from '../../../create/validator';
 import { useImportDatasetDialog } from '../../../providers/import-dataset-dialog-provider.component';
+import { validateProjectName } from '../../../validator';
 import { getRecommendedTaskType, TASK_SELECTION_FORM_ID } from './util';
 
 type ImportTaskSelectionProps = {
